+++ CHANGELOG.md
@@ -21,6 +21,7 @@ New unsafe operations with Postgres
 
 Other
 
+- Added experimental support for lock timeout retries
 - Added `target_sql_mode` option
 - Added error for `change_column_null` with default value with `safe_by_default` option
 - Dropped support for Postgres < 10, MySQL < 5.7, and MariaDB < 10.2